# go-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entities.ApiResponse:
    properties:
      data: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/entities.PaginationResponse'
      success:
        type: boolean
    type: object
...
  entities.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      image:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  entities.CreateProductRequest:
    properties:
      category_id:
        type: string
      description:
        type: string
      image:
        type: string
      images:
        items:
          type: string
        type: array
      name:
        type: string
      price:
        minimum: 0
        type: number
      stock:
        minimum: 0
        type: integer
    required:
    - category_id
    - name
    - price
    type: object
...
  entities.PaginationResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  entities.Product:
    properties:
      category:
        $ref: '#/definitions/entities.Category'
      category_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      image:
        type: string
      images:
        items:
          $ref: '#/definitions/entities.ProductImage'
        type: array
      name:
        type: string
      price:
        type: number
      stock:
        type: integer
      updated_at:
        type: string
    type: object
  entities.ProductImage:
    properties:
      created_at:
        type: string
      id:
        type: string
      image_url:
        type: string
      product_id:
        type: string
      updated_at:
        type: string
    type: object
...
  entities.UpdateProductRequest:
    properties:
      category_id:
        type: string
      description:
        type: string
      image:
        type: string
      images:
        items:
          type: string
        type: array
      name:
        type: string
      price:
        minimum: 0
        type: number
      stock:
        minimum: 0
        type: integer
    type: object
...
host: localhost:4001
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.swagger.io/support
  description: API สำหรับระบบ E-commerce ที่พัฒนาด้วย Go Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go E-commerce API
  version: 1.0.0
paths:
...
  /products:
    get:
      consumes:
      - application/json
      description: ดูสินค้าทั้งหมดพร้อม pagination
      parameters:
      - default: 1
        description: หน้าที่ต้องการ
        in: query
        name: page
        type: integer
      - default: 10
        description: จำนวนรายการต่อหน้า
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entities.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.Product'
                  type: array
                pagination:
                  $ref: '#/definitions/entities.PaginationResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      summary: ดูสินค้าทั้งหมด
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: สร้างสินค้าใหม่ (เฉพาะ Admin)
      parameters:
      - description: ข้อมูลการสร้างสินค้า
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entities.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/entities.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/entities.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      security:
      - BearerAuth: []
      summary: สร้างสินค้า
      tags:
      - Products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: ลบสินค้าตาม ID (เฉพาะ Admin)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      security:
      - BearerAuth: []
      summary: ลบสินค้า
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: ดูรายละเอียดสินค้าตาม ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entities.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/entities.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      summary: ดูสินค้าตาม ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: แก้ไขข้อมูลสินค้า (เฉพาะ Admin)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: ข้อมูลการแก้ไขสินค้า
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entities.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      security:
      - BearerAuth: []
      summary: แก้ไขสินค้า
      tags:
      - Products
  /products/category/{categoryId}:
    get:
      consumes:
      - application/json
      description: ดูสินค้าที่กรองตามหมวดหมู่
      parameters:
      - description: Category ID
        in: path
        name: categoryId
        required: true
        type: string
      - default: 1
        description: หน้าที่ต้องการ
        in: query
        name: page
        type: integer
      - default: 10
        description: จำนวนรายการต่อหน้า
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entities.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.Product'
                  type: array
                pagination:
                  $ref: '#/definitions/entities.PaginationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      summary: ดูสินค้าตามหมวดหมู่
      tags:
      - Products
  /products/search:
    get:
      consumes:
      - application/json
      description: ค้นหาสินค้าตามชื่อ คำอธิบาย หรือเงื่อนไขอื่นๆ
      parameters:
      - description: คำค้นหา
        in: query
        name: search
        type: string
      - description: Category ID
        in: query
        name: category_id
        type: string
      - description: ราคาต่ำสุด
        in: query
        name: min_price
        type: number
      - description: ราคาสูงสุด
        in: query
        name: max_price
        type: number
      - default: 1
        description: หน้าที่ต้องการ
        in: query
        name: page
        type: integer
      - default: 10
        description: จำนวนรายการต่อหน้า
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/entities.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entities.Product'
                  type: array
                pagination:
                  $ref: '#/definitions/entities.PaginationResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ApiResponse'
      summary: ค้นหาสินค้า
      tags:
      - Products
...
securityDefinitions:
  BearerAuth:
    description: 'JWT token สำหรับการยืนยันตัวตน ให้ใส่ token ในรูปแบบ: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"